pyproject.toml
[project]
name = "money-transit-managment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ruff",
    "django",
    "django-rest-framework",
    "django-environ",
    "django-constance",
    "drf-spectacular",
    "gunicorn",
    "django-cors-headers",
    "psycopg2-binary",
    "django-silk",
    "django-mptt",
    "django_filter",
]
